requirements/base.txt
# Base Python Dependencies
# Production requirements

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database & ORM
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
supabase==2.3.0  # Supabase client SDK
postgrest==0.13.0

# Redis/Caching
redis==5.0.1
hiredis==2.3.2

# LLM & ML
torch==2.2.0
transformers==4.36.2
accelerate==0.26.1
peft==0.7.1
bitsandbytes==0.42.0
sentencepiece==0.1.99
protobuf==4.25.2

# Inference
# vllm==0.3.0  # Requires CUDA, skip for Mac local development
together==0.2.9
ollama-python==0.1.2  # For local Ollama integration
openai==1.10.0  # For OpenAI-compatible APIs

# Security & Encryption
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyjwt==2.8.0

# PII Detection
presidio-analyzer==2.2.353
presidio-anonymizer==2.2.353
spacy==3.7.2

# Vector Database & Embeddings
pinecone-client==3.2.0
sentence-transformers==3.0.1
huggingface-hub==0.21.4

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Monitoring & Observability
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-jaeger==1.22.0
opentelemetry-instrumentation-sqlalchemy==0.43b0
opentelemetry-instrumentation-redis==0.43b0
opentelemetry-instrumentation-httpx==0.43b0
python-json-logger==2.0.7

# Data Processing
pandas==2.1.4
numpy==1.26.3
polars==0.19.19

# Monitoring & Logging
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
python-json-logger==2.0.7
loguru==0.7.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tenacity==8.2.3
